requirements.txt
matplotlib
numpy
Pillow
scikit-image
scikit-learn
scipy
sklearn
tensorflow-gpu==1.4.1
tensorflow_compression
